[BUGFIX] Fixing issue with apps mode requirements
